Add footer and pagination on users side list, refactor UserController

+++ app/Http/Controllers/Admin/UserController.php
@@ -6,6 +6,7 @@ use App\Http\Controllers\Controller;
 //use Illuminate\Http\Request;
 use App\Models\User;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Storage;
 
 
 class UserController extends Controller
@@ -19,13 +20,28 @@ class UserController extends Controller
             abort(403);
         }
 
-        $users = User::latest()->paginate();
+        $users = User::latest()->paginate(10);
 
         return view('users.index', compact('users'));
     }
 
     public function destroy(User $user)
     {
+        if ($user->id === Auth::id()) {
+            return back()->with('error', 'You cannot delete yourself');
+        }
+
+        if ($user->isAdmin()) {
+            return back()->with('error', 'You cannot delete another admin');
+        }
+
+        $userAva = $user->avatar;
+        if ($userAva && $userAva !== 'avatars/av_def.png') {
+            Storage::disk('public')->delete($userAva);
+            $user->update(['avatar' => 'avatars/av_def.png']);
+            
+        }
+
         $user->delete();
 
         return back()->with('success', "$user->name was deleted");
